# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/recommendations/{userId}":{"get":{"operationId":"RecommendationController_getRecommendationsForUser","summary":"Get recommendations for a user","parameters":[{"name":"userId","required":true,"in":"path","description":"The ID of the user","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved recommendations."}},"tags":["Recommendations"]}},"/recommendations/generate/{userId}":{"get":{"operationId":"RecommendationController_generateRecommendations","summary":"Generate new recommendations for a user","parameters":[{"name":"userId","required":true,"in":"path","description":"The ID of the user","schema":{"type":"string"}}],"responses":{"201":{"description":"Recommendations successfully generated."}},"tags":["Recommendations"]}},"/items":{"post":{"operationId":"ItemController_createItem","summary":"Create a new item","parameters":[],"requestBody":{"required":true,"description":"The payload for creating a new item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateItemDto"},"examples":{"example1":{"summary":"Sample item payload","value":{"title":"Product Title","description":"This is a description of the product.","type":"product","metadata":"{\"price\": 100, \"color\": \"red\"}"}}}}}},"responses":{"201":{"description":"Item successfully created."}},"tags":["Items"]},"get":{"operationId":"ItemController_getItems","summary":"Retrieve all items","parameters":[],"responses":{"200":{"description":"Successfully retrieved items."}},"tags":["Items"]}},"/users":{"post":{"operationId":"UserController_createUser","summary":"Create a new user","parameters":[],"requestBody":{"required":true,"description":"Payload to create a new user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"},"examples":{"example1":{"summary":"Sample user payload","value":{"name":"John Doe","email":"johndoe@example.com","age":25,"gender":"Male","location":"New York"}}}}}},"responses":{"201":{"description":"User successfully created."}},"tags":["Users"]},"get":{"operationId":"UserController_getAllUsers","summary":"Get all users","parameters":[],"responses":{"200":{"description":"List of all users."}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UserController_getUserById","summary":"Get user by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User data retrieved."}},"tags":["Users"]}},"/interactions":{"post":{"operationId":"UserInteractionController_logInteraction","summary":"Log a user interaction","parameters":[],"requestBody":{"required":true,"description":"The payload for logging a user interaction","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserInteractionDto"},"examples":{"example1":{"summary":"Sample interaction payload","value":{"userId":"7c0e9d85-b499-435b-8a92-5c9d17cc56a2","itemId":"1d2e9f85-d499-43e5-bd65-3f9b17dd25a2","interactionType":"view","rating":5}}}}}},"responses":{"201":{"description":"User interaction successfully logged."}},"tags":["User Interactions"]}},"/interactions/{userId}":{"get":{"operationId":"UserInteractionController_getUserInteractions","summary":"Get user interactions by userId","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User interactions retrieved successfully."}},"tags":["User Interactions"]}},"/interactions/like":{"post":{"operationId":"UserInteractionController_likeItem","summary":"Like or Dislike an item","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeItemDto"}}}},"responses":{"201":{"description":""}},"tags":["User Interactions"]}},"/interactions/rate":{"post":{"operationId":"UserInteractionController_rateItem","summary":"Rate an item","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RateItemDto"}}}},"responses":{"201":{"description":""}},"tags":["User Interactions"]}}},"info":{"title":"AI Recommendation System","description":"API documentation for the AI Recommendation System","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateItemDto":{"type":"object","properties":{"title":{"type":"string","example":"Product Title"},"description":{"type":"string","example":"This is a description of the product."},"type":{"type":"string","example":"product","enum":["product","video","content"]},"metadata":{"type":"string","example":"{\"price\": 100, \"color\": \"red\"}"}},"required":["title","description","type","metadata"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"johndoe@example.com"},"age":{"type":"number","example":25},"gender":{"type":"string","example":"Male"},"location":{"type":"string","example":"New York"}},"required":["name","email","age","gender","location"]},"CreateUserInteractionDto":{"type":"object","properties":{"userId":{"type":"string","description":"The ID of the user","example":"7c0e9d85-b499-435b-8a92-5c9d17cc56a2"},"itemId":{"type":"string","description":"The ID of the item","example":"1d2e9f85-d499-43e5-bd65-3f9b17dd25a2"},"interactionType":{"type":"string","description":"The type of interaction","example":"view","enum":["view","click","purchase","like","rating"]},"rating":{"type":"number","description":"The rating given by the user (optional)","example":5},"like":{"type":"boolean","description":"Whether the item is liked by the user (optional)","example":true}},"required":["userId","itemId","interactionType"]},"LikeItemDto":{"type":"object","properties":{"userId":{"type":"string","description":"The ID of the user","example":"7c0e9d85-b499-435b-8a92-5c9d17cc56a2"},"itemId":{"type":"string","description":"The ID of the item","example":"1d2e9f85-d499-43e5-bd65-3f9b17dd25a2"},"like":{"type":"boolean","description":"True for like, false for dislike","example":true}},"required":["userId","itemId","like"]},"RateItemDto":{"type":"object","properties":{"userId":{"type":"string","description":"The ID of the user","example":"7c0e9d85-b499-435b-8a92-5c9d17cc56a2"},"itemId":{"type":"string","description":"The ID of the item","example":"1d2e9f85-d499-43e5-bd65-3f9b17dd25a2"},"rating":{"type":"number","description":"Rating value between 1 and 5","example":4}},"required":["userId","itemId","rating"]}}}}